requirements.txt
streamlit
supabase
google-generativeai
pypdf
toml
requests
yfinance
pandas
pandas_ta
numpy
matplotlib
ta
aiohttp